fix AccountController

+++ ShopSecondHand/Data/ResponseModels/ProductResponse/GetProductResponse.cs
@@ -7,7 +7,7 @@ namespace ShopSecondHand.Data.ResponseModels.ProductResponse
         public Guid Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int? Price { get; set; }
+        public float? Price { get; set; }
         public int? Quantity { get; set; }
         public Guid? CategoryId { get; set; }
     }
